Move ZDP command handlers into own file
