+++ de_web.pro
@@ -140,7 +140,8 @@ HEADERS  = bindings.h \
            thermostat_ui_configuration.h \
            tuya.h \
            utils/utils.h \
-           websocket_server.h
+           websocket_server.h \
+           zdp/zdp_handlers.h
 
 SOURCES  = air_quality.cpp \
            aps_controller_wrapper.cpp \
@@ -215,7 +216,8 @@ SOURCES  = air_quality.cpp \
            zcl_tasks.cpp \
            window_covering.cpp \
            websocket_server.cpp \
-           xmas.cpp
+           xmas.cpp \
+           zdp/zdp_handlers.cpp
 
 win32 {
